fix error of constructors with generator params

+++ test/runtests.jl
@@ -3,5 +3,5 @@ using Test
 
 g = GeneralGraph{Int}(() -> 1, "blogs", "blogs.txt")
 g = NormalUnweightedGraph("blogs", "blogs.txt")
-g = GeneralDiGraph{Int}(() -> 1, "hepth", "hepth.txt")
+g = GeneralDiGraph{Int8}(() -> Int8(1), "hepth", "hepth.txt")
 g = NormalUnweightedDiGraph("hepth", "hepth.txt")
